Disable web security, set storage quota
I saw these settings in an example that BrightSign support sent me.

`storage_quota: "2000000000"` is a bit sketchy because their doc
says that

> If the storage path is specified without a storage quota, Chromium defaults to reserving 1GB plus 10% of the total size of the storage device.

Which means that 2GB quota will be less than a default quota on
a 32GB disk (4 GB) but it benefits small 4 GB disks (which defaults to 1.3 GB).